+++ src/main/java/com/braintreegateway/Configuration.java
@@ -1,12 +1,12 @@
 package com.braintreegateway;
 
-import com.braintreegateway.exceptions.ConfigurationException;
-import com.braintreegateway.util.ClientLibraryProperties;
-
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
+import com.braintreegateway.exceptions.ConfigurationException;
+import com.braintreegateway.util.ClientLibraryProperties;
 
 public class Configuration {
     private Environment environment;
@@ -27,6 +27,7 @@ public class Configuration {
     }
 
     public static final String VERSION = new ClientLibraryProperties().version();
+    public static final String GRAPHQL_API_VERSION = "2018-05-21";
 
     public static String apiVersion() {
         return "4";
@@ -110,6 +111,10 @@ public class Configuration {
 
     public String getBaseURL() {
         return environment.baseURL;
+    }
+
+    public String getGraphQLURL() {
+        return environment.graphQLURL;
     }
 
     public Boolean usesProxy() {