+++ src/main/java/com/braintreegateway/Environment.java
@@ -7,26 +7,32 @@ import java.util.Arrays;
  */
 public class Environment {
     /** For Braintree internal development. */
-    public static final Environment DEVELOPMENT = new Environment(developmentBaseURL() + ":" + developmentPort(), "http://auth.venmo.dev:9292", new String[] {}, "development");
-    public static final Environment QA = new Environment("https://gateway.qa.braintreepayments.com:443", "https://auth.qa.venmo.com", new String[] {"ssl/api_braintreegateway_com.ca.crt"}, "qa");
+    public static final Environment DEVELOPMENT = new Environment(developmentBaseURL() + ":" + developmentPort(), "http://auth.venmo.dev:9292", new String[]{"ssl/atmosphere.server.crt"}, "development", developmentGraphQLURL());
+    public static final Environment QA = new Environment("https://gateway.qa.braintreepayments.com:443", "https://auth.qa.venmo.com", new String[]{"ssl/api_braintreegateway_com.ca.crt", "ssl/payments_braintreeapi_com.ca.crt"}, "qa", "https://payments-qa.dev.braintree-api.com/graphql");
 
     /** For production. */
-    public static final Environment PRODUCTION = new Environment("https://api.braintreegateway.com:443", "https://auth.venmo.com", new String[] {"ssl/api_braintreegateway_com.ca.crt"}, "production");
+    public static final Environment PRODUCTION = new Environment("https://api.braintreegateway.com:443", "https://auth.venmo.com", new String[]{"ssl/api_braintreegateway_com.ca.crt", "ssl/payments_braintreeapi_com.ca.crt"}, "production", "https://payments.braintree-api.com/graphql");
 
     /** For merchants to use during their development and testing. */
-    public static final Environment SANDBOX = new Environment("https://api.sandbox.braintreegateway.com:443", "https://auth.sandbox.venmo.com", new String[] {"ssl/api_braintreegateway_com.ca.crt"}, "sandbox");
+    public static final Environment SANDBOX = new Environment("https://api.sandbox.braintreegateway.com:443", "https://auth.sandbox.venmo.com", new String[]{"ssl/api_braintreegateway_com.ca.crt", "ssl/payments_braintreeapi_com.ca.crt"}, "sandbox", "https://payments.sandbox.braintree-api.com/graphql");
 
     private String environmentName;
 
     public final String baseURL;
+    public final String graphQLURL;
     public final String authURL;
     public final String[] certificateFilenames;
 
     public Environment(String baseURL, String authURL, String[] certificateFilenames, String environmentName) {
+        this(baseURL, authURL, certificateFilenames, environmentName, null);
+    }
+
+    public Environment(String baseURL, String authURL, String[] certificateFilenames, String environmentName, String graphQLURL) {
         this.baseURL = baseURL;
         this.authURL = authURL;
         this.certificateFilenames = Arrays.copyOf(certificateFilenames, certificateFilenames.length);
         this.environmentName = environmentName;
+        this.graphQLURL = graphQLURL;
     }
 
     public static Environment parseEnvironment(String environment) {
@@ -40,6 +46,14 @@ public class Environment {
             return PRODUCTION;
         } else {
             throw new IllegalArgumentException("Unknown environment: " + environment);
+        }
+    }
+
+    private static String developmentGraphQLURL() {
+        if (System.getenv().get("GRAPHQL_URL") != null) {
+            return System.getenv().get("GRAPHQL_URL");
+        } else {
+            return "https://atmosphere.bt.local:8080/graphql";
         }
     }