+++ src/test/java/com/braintreegateway/integrationtest/IntegrationTest.java
@@ -1,10 +1,30 @@
 package com.braintreegateway.integrationtest;
 
-import com.braintreegateway.*;
-
+import java.io.ByteArrayOutputStream;
+import java.io.OutputStream;
+import java.util.Calendar;
+import java.util.logging.Handler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.Calendar;
+import java.util.logging.StreamHandler;
+
+import com.braintreegateway.BraintreeGateway;
+import com.braintreegateway.Configuration;
+import com.braintreegateway.Environment;
+import com.braintreegateway.ResourceCollection;
+import com.braintreegateway.SandboxValues;
+import com.braintreegateway.Transaction;
+import com.braintreegateway.TransactionRequest;
+import com.braintreegateway.TransactionSearchRequest;
+
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.OutputStream;
+import java.net.URL;
+import java.util.logging.Handler;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import java.util.logging.StreamHandler;
 
 import org.junit.Before;
 
@@ -13,6 +33,10 @@ public class IntegrationTest {
     protected BraintreeGateway gateway;
     protected Transaction disputedTransaction;
 
+    protected static OutputStream logCapturingStream;
+    protected static StreamHandler customLogHandler;
+
+
     @Before
     public void createGateway() {
         this.gateway = new BraintreeGateway(
@@ -20,6 +44,15 @@ public class IntegrationTest {
             "integration_merchant_id",
             "integration_public_key",
             "integration_private_key"
+        );
+    }
+
+    public void createAdvancedFraudMerchantGateway() {
+        this.gateway = new BraintreeGateway(
+            Environment.DEVELOPMENT,
+            "advanced_fraud_integration_merchant_id",
+            "advanced_fraud_integration_public_key",
+            "advanced_fraud_integration_private_key"
         );
     }
 
@@ -78,5 +111,22 @@ public class IntegrationTest {
         this.gateway.getConfiguration().setLogger(Logger.getLogger("null"));
         this.gateway.getConfiguration().getLogger().setLevel(Level.INFO);
         this.gateway.getConfiguration().getLogger().setUseParentHandlers(false);
+    }
+
+    protected void attachLogCapturer()
+    {
+        Configuration configuration = this.gateway.getConfiguration();
+        Logger logger = configuration.getLogger();
+        logCapturingStream = new ByteArrayOutputStream();
+        Handler[] handlers = logger.getParent().getHandlers();
+        customLogHandler = new StreamHandler(logCapturingStream, handlers[0].getFormatter());
+        customLogHandler.setLevel(Level.FINE);
+        logger.addHandler(customLogHandler);
+    }
+
+    protected String getTestCapturedLog()
+    {
+      customLogHandler.flush();
+      return logCapturingStream.toString();
     }
 }